backend/requirements.txt
fastapi
uvicorn
pymongo
pyjwt
passlib[bcrypt]